% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_transactions.R
\name{earned}
\alias{earned}
\title{Add income}
\usage{
earned(date = "default", value = "default", account = "default",
  situation = "default", source = "default", obs = "default")
}
\description{
Add an income to current database
}

