% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_transactions.R
\name{erase_transaction}
\alias{erase_transaction}
\title{Erase transaction}
\usage{
erase_transaction(n = 1, type = "expenses")
}
\arguments{
\item{n}{A vector with the positions of the transactions to be excluded, 1 being the most recent one.}

\item{type}{Type of transaction, either "expenses" or "incomes".}
}
\description{
Erase a transaction from the record.
}

