% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_transactions.R
\name{read_expenses}
\alias{read_expenses}
\title{Read expenses}
\usage{
read_expenses()
}
\value{
Dataframe with record of expenses of the current database.
}
\description{
Read record of expenses of the current database.
}

