% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_transactions.R
\name{read_incomes}
\alias{read_incomes}
\title{Read incomes}
\usage{
read_incomes()
}
\value{
Dataframe with record of incomes of the current database.
}
\description{
Read record of incomes of the current database.
}

