% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_transactions.R
\name{spent}
\alias{spent}
\title{Add expense}
\usage{
spent(date = "default", value = "default", account = "default",
  situation = "default", type = "default", subtype = "default",
  obs = "default")
}
\description{
Add an expense to current database
}

