% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_accounts.R
\name{view_accounts}
\alias{view_accounts}
\title{View current accounts}
\usage{
view_accounts()
}
\value{
A list in which the first element is a vector of the names, and the second one is a vector of the currencies of each of the accounts in the current database.
}
\description{
Returns a list with the names and currencies of the accounts in the current database.
}

